EOS-21429: moved ha proxy config code to S3HAproxy class Signed-off-by: Vimal Raghwani <vimal.raghwani@seagate.com>
